Remove foundation in some places
