+++ Sources/AdaEngine/Scene/Serialization/SceneSerializer.swift
@@ -5,7 +5,6 @@
 //  Created by v.prusakov on 5/6/22.
 //
 
-import Foundation
 import Yams
 
 struct SystemRepresentation: Codable {
